Allow caching to be controlled in the config file (and in a more fine-grained way)
Add cache_workspaces and cache_windows to the config to individually control what gets cached, but still allow the command-line flag to override the settings and disable the cache entirely.
Why?
Caching window positions can be a bit unreliable for things like browser windows (which will all have the same WM_CLASS and window title on start-up so may not match between runs/logins when cortile encounters the window).